Moved wokibrain application to root folder
